+++ src/libr/include/r_asm.h
@@ -39,62 +39,62 @@ enum {
 };
 
 struct r_asm_t {
-	u32  arch;
-	u32  bits;
-	u32  big_endian;
-	u32  syntax;
-	u32  parser;
+	int  arch;
+	int  bits;
+	int  big_endian;
+	int  syntax;
+	int  parser;
 	u64  pc;
 	char buf_asm[256];
 	char buf_hex[256];
 	char buf_err[256];
 	void *aux;
-	u32  (*r_asm_disasm)(struct r_asm_t *a, u8 *buf, u32 len);
-	u32  (*r_asm_asm)(struct r_asm_t *a, char *buf);
-	u32  (*r_asm_parse)(struct r_asm_t *a);
-	u32  (*r_asm_parse_cb)(struct r_asm_t *a);
+	int  (*r_asm_disasm)(struct r_asm_t *a, u8 *buf, u64 len);
+	int  (*r_asm_asm)(struct r_asm_t *a, char *buf);
+	int  (*r_asm_parse)(struct r_asm_t *a);
+	int  (*r_asm_parse_cb)(struct r_asm_t *a);
 };
 
 /* asm.c */
 int r_asm_init(struct r_asm_t *a);
 struct r_asm_t *r_asm_new();
 void r_asm_free(struct r_asm_t *a);
-int r_asm_set_arch(struct r_asm_t *a, u32 arch);
-int r_asm_set_bits(struct r_asm_t *a, u32 bits);
-int r_asm_set_big_endian(struct r_asm_t *a, u32 boolean);
-int r_asm_set_syntax(struct r_asm_t *a, u32 syntax);
-int r_asm_set_parser(struct r_asm_t *a, u32 parser, 
-		u32 (*cb)(struct r_asm_t *a), void *aux);
+int r_asm_set_arch(struct r_asm_t *a, int arch);
+int r_asm_set_bits(struct r_asm_t *a, int bits);
+int r_asm_set_big_endian(struct r_asm_t *a, int boolean);
+int r_asm_set_syntax(struct r_asm_t *a, int syntax);
+int r_asm_set_parser(struct r_asm_t *a, int parser, 
+		int (*cb)(struct r_asm_t *a), void *aux);
 int r_asm_set_pc(struct r_asm_t *a, u64 pc);
-u32 r_asm_disasm(struct r_asm_t *a, u8 *buf, u32 len);
-u32 r_asm_asm(struct r_asm_t *a, char *buf);
-u32 r_asm_parse(struct r_asm_t *a);
+int r_asm_disasm(struct r_asm_t *a, u8 *buf, u64 len);
+int r_asm_asm(struct r_asm_t *a, char *buf);
+int r_asm_parse(struct r_asm_t *a);
 
 /* arch/x86/asm.c */
-u32 r_asm_x86_disasm(struct r_asm_t *a, u8 *buf, u32 len);
-u32 r_asm_x86_asm(struct r_asm_t *a, char *buf);
+int r_asm_x86_disasm(struct r_asm_t *a, u8 *buf, u64 len);
+int r_asm_x86_asm(struct r_asm_t *a, char *buf);
 /* arch/x86/pseudo.c */
-u32 r_asm_x86_pseudo(struct r_asm_t *a);
+int r_asm_x86_pseudo(struct r_asm_t *a);
 /* arch/x86/realloc.c */
 struct r_asm_realloc_t {
 	u64 offset;
 	u64 delta;
 	char str[256];
 };
-u32 r_asm_x86_realloc(struct r_asm_t *a);
+int r_asm_x86_realloc(struct r_asm_t *a);
 
 /* arch/arm/asm.c */
-u32 r_asm_arm_disasm(struct r_asm_t *a, u8 *buf, u32 len);
+int r_asm_arm_disasm(struct r_asm_t *a, u8 *buf, u64 len);
 
 /* arch/mips/asm.c */
-u32 r_asm_mips_disasm(struct r_asm_t *a, u8 *buf, u32 len);
+int r_asm_mips_disasm(struct r_asm_t *a, u8 *buf, u64 len);
 
 /* arch/sparc/asm.c */
-u32 r_asm_sparc_disasm(struct r_asm_t *a, u8 *buf, u32 len);
+int r_asm_sparc_disasm(struct r_asm_t *a, u8 *buf, u64 len);
 
 /* arch/ppc/asm.c */
-u32 r_asm_ppc_disasm(struct r_asm_t *a, u8 *buf, u32 len);
+int r_asm_ppc_disasm(struct r_asm_t *a, u8 *buf, u64 len);
 
 /* arch/bf/asm.c */
-u32 r_asm_bf_disasm(struct r_asm_t *a, u8 *buf, u32 len);
+int r_asm_bf_disasm(struct r_asm_t *a, u8 *buf, u64 len);
 #endif
